requirements.txt
flask==2.0.1
gunicorn==20.1.0
google-cloud-vision==2.4.2
supabase==1.0.3
python-dotenv==0.19.0
cryptography==41.0.7
Werkzeug==2.0.3
uuid==1.30
datetime
requests==2.31.0
Flask-CORS==4.0.0
